requirements.txt
slack_bolt==1.25.0
apscheduler==3.11.0
python-dotenv==1.1.1
google-generativeai==0.8.5
google-auth==2.41.1
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.183.0
certifi==2025.8.3
urllib3==2.5.0
requests==2.32.5
